mono_depth/requirements.txt
torch==1.9.0
torchvision==0.10.0
imageio
scipy
argparse
Pillow
sacred
easydict
pymongo
pyyaml
